Switch to Apache License, Version 2.0
I considered using GPLv3 or another copyleft license,
but I don't want people to be afraid to use this code
in their projects. In my opinion, somebody who finds
my code useful will probably contribute their changes
back to this project because doing so is usually in
the best interest of the contributor.

Signed-off-by: Paul Morgan <jumanjiman@gmail.com>