Improve and document the bundler setup
This commit affects how bundler treats this gem
when required as a dependency from another gem.
We can use a library version of bundle configuration
as long as we require bundle/setup first.

Once we've done our bundler setup, we can freely use any
"sub-libs" of this gem anywhere in this gem, including the
Rakefile.

Before this commit:
* `rake -T' fails to load deps that are required within modules
* `bundle exec rake -T' works

After this commit:
* `rake -T' works as expected
* `bundle exec rake -T' works

Signed-off-by: Paul Morgan <jumanjiman@gmail.com>