repo config: add minimal settings
Provide a minimal configuration that is applied to the local
git repo, overriding user-default personal preferences. This
is not a dumping ground; it's intended to provide consistency.

The upstream remote should use either https or git protocol
(not ssh) to ensure that continuous integration tools do not
need special configuration. This is a public repo; strong
authentication should not be needed for CI.

New unit test:

    .repo/config
      does not use ssh for upstream remote

Signed-off-by: Paul Morgan <jumanjiman@gmail.com>